+++ lib/AutoBackend.Sdk/Controllers/GenericControllerWithComplexKey.4.cs
@@ -1,5 +1,6 @@
+using AutoBackend.Sdk.Data.Repositories;
+using AutoBackend.Sdk.Filters;
 using AutoBackend.Sdk.Models;
-using AutoBackend.Sdk.Storage;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AutoBackend.Sdk.Controllers;
@@ -11,28 +12,37 @@ internal sealed class GenericControllerWithComplexKey<
     TKey2,
     TKey3,
     TKey4
-> : GenericController<TEntity, TFilter>
+> : GenericController<
+    TEntity,
+    TFilter
+>
     where TEntity : class
-    where TFilter : class
+    where TFilter : class, IGenericFilter
+    where TKey1 : notnull
+    where TKey2 : notnull
+    where TKey3 : notnull
+    where TKey4 : notnull
 {
-    private readonly IGenericStorageWithComplexKey<
+    private readonly IGenericRepositoryWithComplexKey<
         TEntity,
+        TFilter,
         TKey1,
         TKey2,
         TKey3,
         TKey4
-    > _genericStorage;
+    > _genericRepository;
 
     public GenericControllerWithComplexKey(
-        IGenericStorageWithComplexKey<
+        IGenericRepositoryWithComplexKey<
             TEntity,
+            TFilter,
             TKey1,
             TKey2,
             TKey3,
             TKey4
-        > genericStorage) : base(genericStorage)
+        > genericRepository) : base(genericRepository)
     {
-        _genericStorage = genericStorage;
+        _genericRepository = genericRepository;
     }
 
     [HttpGet("{key1}/{key2}/{key3}/{key4}")]
@@ -42,7 +52,7 @@ internal sealed class GenericControllerWithComplexKey<
         [FromRoute] TKey3 key3,
         [FromRoute] TKey4 key4)
     {
-        return ProcessAsync(cancellationToken => _genericStorage.GetByComplexKeyAsync(
+        return ProcessAsync(cancellationToken => _genericRepository.GetByComplexKeyAsync(
             key1,
             key2,
             key3,
@@ -51,14 +61,14 @@ internal sealed class GenericControllerWithComplexKey<
     }
 
     [HttpPost("{key1}/{key2}/{key3}/{key4}")]
-    public Task<ActionResult<GenericControllerResponse<TEntity>>> InsertByComplexKeyAsync(
+    public Task<ActionResult<GenericControllerResponse<TEntity>>> CreateByComplexKeyAsync(
         [FromRoute] TKey1 key1,
         [FromRoute] TKey2 key2,
         [FromRoute] TKey3 key3,
         [FromRoute] TKey4 key4,
         [FromBody] TEntity entity)
     {
-        return ProcessAsync(cancellationToken => _genericStorage.InsertByComplexKeyAsync(
+        return ProcessAsync(cancellationToken => _genericRepository.CreateByComplexKeyAsync(
             key1,
             key2,
             key3,
@@ -75,7 +85,7 @@ internal sealed class GenericControllerWithComplexKey<
         [FromRoute] TKey4 key4,
         [FromBody] TEntity entity)
     {
-        return ProcessAsync(cancellationToken => _genericStorage.UpdateByComplexKeyAsync(
+        return ProcessAsync(cancellationToken => _genericRepository.UpdateByComplexKeyAsync(
             key1,
             key2,
             key3,
@@ -91,7 +101,7 @@ internal sealed class GenericControllerWithComplexKey<
         [FromRoute] TKey3 key3,
         [FromRoute] TKey4 key4)
     {
-        return ProcessAsync(cancellationToken => _genericStorage.DeleteByComplexKeyAsync(
+        return ProcessAsync(cancellationToken => _genericRepository.DeleteByComplexKeyAsync(
             key1,
             key2,
             key3,